added cast to comparation

+++ src/goto-programs/loop_unroll.cpp
@@ -158,7 +158,7 @@ int bounded_loop_unroller::get_loop_bounds(loopst &loop)
     }
 
   int bound = k - k0;
-  if(bound <= 0 || bound > unroll_limit)
+  if(bound <= 0 || (size_t)bound > unroll_limit)
     return 0;
 
   number_of_bounded_loops++;
